+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ### Changed
+- Ocean Watch: updated texts from Partners page. [OW-24](https://vizzuality.atlassian.net/browse/OW-24)
 - optional Redis client locally.
 - widget-editor@2.6.2
 - relocates pages according to Next's dynamic routes. [RW-52](https://vizzuality.atlassian.net/browse/RW-52)